+++ templates/widgets/template.js.ts
@@ -53,10 +53,15 @@ vis.binds["${widgetName}"] = {
 		$('#' + widgetID).html(text);
 
 		// subscribe on updates of value
+		function onChange(e, newVal, oldVal) {
+			$div.find('.template-value').html(newVal);
+		}
 		if (data.oid) {
-			vis.states.bind(data.oid + '.val', function (e, newVal, oldVal) {
-				$div.find('.${widgetName}-value').html(newVal);
-			});
+			vis.states.bind(data.oid + '.val', onChange);
+			//remember bound state that vis can release if didnt needed
+			$div.data('bound', [data.oid + '.val']);
+			//remember onchange handler to release bound states
+			$div.data('bindHandler', onChange);
 		}
 	}
 };